Update Assignment1
Finished part1 and part2 (still need plot for polynomial=20 though).

All of the files are inside the package. main.py can be run for a quick test of all of the parts (main.py imports the package modules).

Make sure housing.csv is in same folder as main.py (default in same directory as the package)